Nested ldap queries (#794)
Added the capability to do ldap queries with nested keys.

Queries are first checked against the compound key and if a value is found it is returned.

If a value is not found, the key is split into segments on the "." character and then the AnyMap is "walked down" to look for a path to the value in sub-maps addressed by the segments of the key. So, if the value of the key "a.b.c" is queried,

1. Split the key into ["a","b","c"]
2. Check the top level map for a sub-map at key "a". If one is found, look in that map for a key "b" with a sub-map, and in that sub-map for a value with key "c". If found, return it.
3. If the top level map does NOT contain a key "a", look for a sub-map at key "a.b". If one is found, look in that sub-map for a value with a key of "c". If one is found, return it.
4. continue this algorithm down as many levels as there are segments until either the value for the last segment is found (in which case it is returned), or one is not found, in which case the lookup fails.

Additional changes to cherry-picked commit:
* convert from C++-17 to C++-14 with Abseil (include more parts of Abseil)
Signed-off-by: Ingmar Sittl <ingmar.sittl@elektrobit.com>